HydraAuction.Delegate.Interface

Documentation

data DelegateResponse

Constructors

CurrentDelegateState ResponseReason DelegateState

RequestIgnored RequestIgnoredReason

AuctionSet AuctionTerms

data DelegateState

This is an important type. It should encode all Delegate server state required for its logic It should be (eventually) same for all servers, synced by Hydra events Client should have up-to date state, gotten by pushed responses.

Constructors

NotInitialized

Initialized HeadId InitializedState

data InitializedState

Constructors

AwaitingCommits

Fields

stangingBidWasCommited :: Bool

Open OpenHeadUtxo (Maybe AuctionTerms)

Closed

Finalized

AbortRequested AbortReason

Aborted

data InitializedStateKind

Constructors

AwaitingCommitsKind

OpenKind

ClosedKind

FinalizedKind

AbortRequestedKind

AbortedKind

initializedStateKind :: InitializedState -> InitializedStateKind

data RequestIgnoredReason

Constructors

IncorrectRequestData IncorrectRequestDataReason

WrongDelegateState DelegateState

data ResponseReason

Constructors

WasQueried

Updated

data FrontendRequest

Constructors

QueryCurrentDelegateState

CommitStandingBid

Fields

auctionTerms :: AuctionTerms

utxoToCommit :: (TxIn, TxOut CtxUTxO)

NewBid

Fields

auctionTerms :: AuctionTerms

datum :: StandingBidDatum

data IncorrectRequestDataReason

Constructors

AuctionTermsAreInvalidOrNotMatchingHead

InvalidBidTerms

TxIdDoesNotExist

data ImposibleEvent

Constructors

IncorrectCommit

IncorrectStandingBidUtxoOnL2

IncorrectHydraEvent

OnChainInvariantBreaks

data OpenHeadUtxo

Constructors

MkOpenHeadUtxo

Fields

standingBidTerms :: Maybe BidTerms

standingBidUtxo :: (TxIn, TxOut CtxUTxO)

collateralUtxo :: (TxIn, TxOut CtxUTxO)

data MissingPrerequisite

Constructors

AdaForCommit

HydraInit

data AbortReason

Constructors

ImpossibleHappened ImposibleEvent

RequiredHydraRequestFailed

PrerequisiteMissing MissingPrerequisite

NobodyCommitedInTime

initialState :: DelegateState

isFinalState :: DelegateState -> Bool

wasOpened :: DelegateState -> Bool