HydraAuction.OnChain.Common

Documentation

minAuctionFee :: Integer

validAuctionTerms :: AuctionTerms -> Bool

checkVoucherExpiredOrLater :: AuctionTerms -> TxInfo -> Bool

checkInterval :: AuctionTerms -> AuctionStage -> TxInfo -> Bool

secondsLeftInInterval :: POSIXTime -> Interval POSIXTime -> Maybe Integer

Given a POSIXTime, and an Interval this function computes the truncated difference in seconds to the UpperBound of the passed Interval. If the Interval does not have a finite UpperBound, or if the given time is past the finite UpperBound the function will return Nothing. Note that this function is not and should not be used on-chain

stageToInterval :: AuctionTerms -> AuctionStage -> Interval POSIXTime

strictTo :: a -> Interval a