HydraAuction.Types

Documentation

isStarted :: AuctionState -> Bool

data BidTerms

Constructors

BidTerms

Fields

bidderPKH :: !PubKeyHash

PubKeyHash of whoever submitted the bid?

bidderVK :: BuiltinByteString

Verification Key of whoever submitted the bid

bidAmount :: !Natural

Which amount did the bidder set to buy the auction lot?

bidderSignature :: BuiltinByteString

Represents the signed payload by the bidder

sellerSignature :: BuiltinByteString

Represents the signed payload by the seller

data BidDepositDatum

Constructors

BidDepositDatum

Fields

bidDepositBidder :: !PubKeyHash

Which bidder made this deposit?

bidDepositVoucherCS :: !VoucherCS

data BidDepositRedeemer

Constructors

LosingBidder

WinningBidder

SellerClaimsDeposit

CleanupDeposit

newtype StandingBidState

Constructors

StandingBidState

Fields

standingBid :: Maybe BidTerms

data StandingBidDatum

Constructors

StandingBidDatum

Fields

standingBidState :: !StandingBidState

standingBidVoucherCS :: !VoucherCS

data AuctionTerms

Constructors

AuctionTerms

Fields

auctionLot :: !AssetClass

What is being sold at the auction?

sellerPKH :: !PubKeyHash

Who is selling it?

sellerVK :: !BuiltinByteString

Verification key of the seller

hydraHeadId :: !CurrencySymbol

Which Hydra Head is authorized to host the bidding for this auction?

delegates :: ![PubKeyHash]

Who is running the Hydra Head where bidding occurs?

biddingStart :: !POSIXTime

Auction lifecycle times.

biddingEnd :: !POSIXTime

voucherExpiry :: !POSIXTime

cleanup :: !POSIXTime

auctionFeePerDelegate :: !Natural

Each delegate will receive this fee portion from the proceeds of the auction, when the auction lot is purchased or reclaimed.

startingBid :: !Natural

The auction lot cannot be sold for less than this bid price.

minimumBidIncrement :: !Natural

A new bid can only supersede the standing bid if it is larger by this increment.

utxoNonce :: !TxOutRef

The seller consumed this utxo input in the transaction that announced this auction, to provide the auction lot to the auction.

data AuctionState

Constructors

Announced

BiddingStarted

data AuctionEscrowDatum

Constructors

AuctionEscrowDatum

Fields

auctionState :: !AuctionState

auctionVoucherCS :: !VoucherCS

data EscrowRedeemer

Constructors

StartBidding

SellerReclaims

BidderBuys

data StandingBidRedeemer

Constructors

MoveToHydra

NewBid

Cleanup

type FeeEscrowDatum = ()

data FeeEscrowRedeemer

Constructors

DistributeFees

data VoucherForgingRedeemer

Constructors

MintVoucher

BurnVoucher

calculateTotalFee :: AuctionTerms -> Integer

data AuctionStage

Constructors

AnnouncedStage

BiddingStartedStage

BiddingEndedStage

VoucherExpiredStage

CleanupStage

auctionStages :: [AuctionStage]