HydraAuctionUtils.L1.Runner

Documentation

data HydraAuctionLog

Constructors

FromHydraAuction !String

data EndToEndLog

Constructors

FromCardanoNode NodeLog

FromFaucet FaucetLog

data NodeLog

Constructors

MsgNodeCmdSpec Text

MsgCLI [Text]

MsgCLIStatus Text Text

MsgCLIRetry Text

MsgCLIRetryResult Text Int

MsgNodeStarting

Fields

stateDirectory :: FilePath

MsgSocketIsReady FilePath

MsgSynchronizing

Fields

percentDone :: Centi

MsgNodeIsReady

data L1Runner a

HydraAuction specific L1 computation executor.

executeL1Runner :: ExecutionContext -> L1Runner a -> IO a

executeL1RunnerWithNodeAs :: forall x. RunningNode -> Actor -> WithActorT L1Runner x -> IO x

executeL1RunnerWithNode :: forall x. RunningNode -> L1Runner x -> IO x

executeTestL1Runner :: L1Runner () -> IO ()

Executes a test runner using a temporary directory as the StateDirectory.

dockerNode :: RunningNode

newtype StateDirectory

FilePath used to store the running node data.

Constructors

MkStateDirectory

Fields

stateDirectory :: FilePath

data ExecutionContext

Execution context holding the current tracer, as well as the running node.

Constructors

MkExecutionContext

Fields

tracer :: Tracer IO HydraAuctionLog

node :: RunningNode

withActor :: forall m x. Monad m => Actor -> WithActorT m x -> m x

fileTracer :: StateDirectory -> IO (Tracer IO HydraAuctionLog)

Filter tracer which logs into a `test.log` file within the given StateDirectory.

toSlotNo :: MonadBlockchainParams m => POSIXTime -> m SlotNo

initWallet :: forall m. (MonadIO m, MonadFail m, MonadTrace m, MonadCardanoClient m) => Lovelace -> Actor -> m ()

Initiates the actor's wallet using the prescribed amount of faucet Lovelace.

stdoutOrNullTracer :: Verbosity -> IO (Tracer IO HydraAuctionLog)

Stdout or null tracer depending the given verbosity level.