HydraAuctionUtils.Tracing

tracing transformer

newtype TracerT t m a

a transformer that provides a contravariant tracer

Constructors

TracerT

Fields

runTracerT :: ReaderT (Tracer m t) m a

runWithTracer :: forall a t m n. (forall b. n b -> m b) -> Tracer n t -> TracerT t m a -> m a

run a tracer by first applying a natural transformation to the tracer

runWithTracer' :: forall a t m. Tracer m t -> TracerT t m a -> m a

like runWithTracer but using id as the natural transformation

askTracer :: forall t m. Monad m => TracerT t m (Tracer m t)

ask the Tracer transformer for its Tracer, tthis just wraps the MonadReader instance method

tracing type class

class MonadTracer t m | m -> t where

a monad that one can trace in the typeclass takes as argument the type that serves as the trace and the monad that you can trace in

Methods

trace :: t -> m ()

trace a message in the monad